Worklfow index -> FastAPI.
